chore(coin): start removing some proxy files and update LLM import in accordance
