+++ apps/ledger-live-mobile/src/components/CurrencyUnitValue.tsx
@@ -1,9 +1,9 @@
 import React, { useContext } from "react";
 import { formatCurrencyUnit } from "@ledgerhq/live-common/currencies/index";
+import type { formatCurrencyUnitOptions } from "@ledgerhq/live-common/currencies/index";
 import { Unit } from "@ledgerhq/types-cryptoassets";
 import { useSelector } from "react-redux";
 import { BigNumber } from "bignumber.js";
-import { formatCurrencyUnitOptions } from "@ledgerhq/live-common/currencies/formatCurrencyUnit";
 import { useLocale } from "../context/Locale";
 import DiscreetModeContext from "../context/DiscreetModeContext";
 import { discreetModeSelector } from "../reducers/settings";
