+++ apps/ledger-live-mobile/src/families/tron/AccountBalanceSummaryFooter.tsx
@@ -2,8 +2,10 @@ import React, { useCallback, useMemo, useState } from "react";
 import { ScrollView } from "react-native";
 import { useSelector } from "react-redux";
 import { useTranslation } from "react-i18next";
-import { getCryptoCurrencyById } from "@ledgerhq/live-common/currencies/index";
-import { toLocaleString } from "@ledgerhq/live-common/currencies/BigNumberToLocaleString";
+import {
+  getCryptoCurrencyById,
+  toLocaleString,
+} from "@ledgerhq/live-common/currencies/index";
 import { getAccountUnit } from "@ledgerhq/live-common/account/helpers";
 import { getCryptoCurrencyIcon } from "@ledgerhq/live-common/reactNative";
 import type { Account } from "@ledgerhq/types-live";
