+++ apps/ledger-live-mobile/src/screens/Portfolio/PortfolioHistory.tsx
@@ -7,7 +7,7 @@ import {
 import { useSelector } from "react-redux";
 import { useFocusEffect } from "@react-navigation/native";
 import type { DailyOperationsSection, Operation } from "@ledgerhq/types-live";
-import { groupAccountsOperationsByDay } from "@ledgerhq/live-common/account/groupOperations";
+import { groupAccountsOperationsByDay } from "@ledgerhq/live-common/account/index";
 import { isAccountEmpty } from "@ledgerhq/live-common/account/helpers";
 
 import { Trans } from "react-i18next";
