+++ libs/coin-framework/src/account/index.ts
@@ -1,6 +1,5 @@
 export * from "./accountId";
 export * from "./accountName";
-export * from "./addAccounts";
 export * from "./balanceHistoryCache";
 export * from "./groupOperations";
 export * from "./helpers";
