--- libs/ledger-live-common/src/account/groupOperations.ts
@@ -1,1 +1,0 @@
-export * from "@ledgerhq/coin-framework/account/groupOperations";
