--- libs/ledger-live-common/src/currencies/BigNumberToLocaleString.ts
@@ -1,1 +1,0 @@
-export * from "@ledgerhq/coin-framework/currencies/BigNumberToLocaleString";
