--- libs/ledger-live-common/src/currencies/formatCurrencyUnit.ts
@@ -1,1 +1,0 @@
-export * from "@ledgerhq/coin-framework/currencies/formatCurrencyUnit";
