+++ libs/ledger-live-common/src/families/bitcoin/transaction.ts
@@ -16,7 +16,7 @@ import {
   fromTransactionStatusRawCommon,
   toTransactionCommonRaw,
   toTransactionStatusRawCommon,
-} from "../../transaction/common";
+} from "@ledgerhq/coin-framework/transaction/common";
 import { getAccountUnit } from "../../account";
 import { formatCurrencyUnit } from "../../currencies";
 import type { Account } from "@ledgerhq/types-live";
