+++ libs/ledger-live-common/src/families/evm/transaction.ts
@@ -10,7 +10,7 @@ import {
   fromTransactionStatusRawCommon as fromTransactionStatusRaw,
   toTransactionCommonRaw,
   toTransactionStatusRawCommon as toTransactionStatusRaw,
-} from "../../transaction/common";
+} from "@ledgerhq/coin-framework/transaction/common";
 import type {
   EvmTransactionEIP1559,
   EvmTransactionLegacy,
