--- libs/ledger-live-common/src/nft/nftId.ts
@@ -1,1 +1,0 @@
-export * from "@ledgerhq/coin-framework/nft/nftId";
