--- libs/ledger-live-common/src/transaction/common.ts
@@ -1,1 +1,0 @@
-export * from "@ledgerhq/coin-framework/transaction/common";
